Move diagnostics module into package and update imports
